unified metabase page fn
