--- apps/src/metabase/helpers/sql/utils.ts
@@ -1,8 +1,0 @@
-export const isSQLPageUrl = (url: string) => {
-  return url.includes('/question');
-}
-
-export const isModelPageUrl = (url: string) => {
-  return url.includes('/model');
-}
-
